Combine _ad_base and _nd_filE_hdf into core_io
